+++ actions/update_object.py
@@ -1,3 +1,4 @@
+from infoblox_client import exceptions
 from lib.action import InfobloxBaseAction
 
 __all__ = [
@@ -6,10 +7,15 @@ __all__ = [
 
 
 class UpdateObjectAction(InfobloxBaseAction):
-    def run(self, ref, **kwargs):
-        _keys = list(kwargs.keys())
+    def run(self, ref, payload):
+        _keys = list(payload.keys())
         for k in _keys:
-            if kwargs[k] is None:
-                del kwargs[k]
-        result = self.connection.update_object(ref, kwargs)
-        return result
+            if payload[k] is None:
+                del payload[k]
+
+        try:
+            result = self.connection.update_object(ref, payload)
+        except exceptions.InfobloxException as e:
+            return (False, e)
+
+        return (True, result)