+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Updated returned value format of get_action method from `_ref` to `ref`.
 - Added an action to delete specified A records
 - Added an action to delete specified CNAME records
+- Added an action to update object (infoblox.update_object)
 - Fixed typo of action get_authzone (changed action name from get_authuone to get_authzone)
 
 ## 0.1.0